refactor(tooltip): 修复了导致ci异常的问题
