+++ __tests__/unit/ui/tooltip/index-spec.ts
@@ -100,10 +100,10 @@ describe('tooltip', () => {
       tooltip.hide();
     });
 
-    expect(tooltip.HTMLTooltipElement.getElementsByClassName('tooltip-title')[0].innerHTML).toBe('标题');
+    expect(tooltip.HTMLTooltipElement.getElementsByClassName('tooltip-title')[0]!.innerHTML).toBe('标题');
     let [tooltipList] = tooltip.HTMLTooltipElement.getElementsByClassName('tooltip-list');
     expect(tooltipList.children.length).toBe(2);
-    expect(tooltipList.querySelector('.tooltip-list-item-name').innerHTML).toBe('第一项');
+    expect(tooltipList.querySelector('.tooltip-list-item-name')!.innerHTML).toBe('第一项');
 
     tooltip.update({
       title: '更新了标题',
@@ -129,9 +129,9 @@ describe('tooltip', () => {
       ],
     });
 
-    expect(tooltip.HTMLTooltipElement.getElementsByClassName('tooltip-title')[0].innerHTML).toBe('更新了标题');
+    expect(tooltip.HTMLTooltipElement.getElementsByClassName('tooltip-title')[0]!.innerHTML).toBe('更新了标题');
     [tooltipList] = tooltip.HTMLTooltipElement.getElementsByClassName('tooltip-list');
     expect(tooltipList.children.length).toBe(3);
-    expect(tooltipList.querySelector('.tooltip-list-item-name').innerHTML).toBe('第三项');
+    expect(tooltipList.querySelector('.tooltip-list-item-name')!.innerHTML).toBe('第三项');
   });
 });
